{{cookiecutter.app_name}}/requirements/prod.txt
Flask==1.0.2
flask-cors==3.0.7
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
flask-marshmallow==0.10.0
flask-restful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.1
psycopg2==2.7.7
PyJWT==1.7.1
python-dotenv==0.10.1